refactor: allow for customization of alchemy chains in alchemy client
