+++ lerna.json
@@ -1,6 +1,5 @@
 {
   "$schema": "node_modules/lerna/schemas/lerna-schema.json",
-  "useWorkspaces": true,
   "version": "3.0.0-alpha.0",
   "useNx": true,
   "npmClient": "yarn",
