+++ package.json
@@ -15,6 +15,7 @@
     "generate": "lerna run generate",
     "postgenerate": "yarn lint:write",
     "build": "lerna run build --ignore=alchemy-daapp --ignore=aa-simple-dapp",
+    "build:ci": "yarn build && yarn lint:write",
     "build:examples": "lerna run build",
     "clean": "lerna run clean",
     "test": "lerna run test:run",
@@ -31,7 +32,7 @@
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-import": "^2.27.5",
     "husky": "^8.0.0",
-    "lerna": "^6.6.1",
+    "lerna": "^8.0.2",
     "lint-staged": "^13.2.2",
     "node-fetch": "^3.3.1",
     "prettier": "^2.8.8",
